requirements.txt
streamlit>=1.31.0
PyPDF2>=3.0.0
Pillow>=10.0.0
cohere>=4.32
PyMuPDF>=1.23.0
python-dotenv>=1.0.0